Moving platform added
Here I made a lot of changes. Tried the platform some times. So far is good, just the timeline is a little confusing and the pingpong loop doesn't work.

Here is also an easy way to apply Waypoint to platforms for later, but this does not work with the controller...